FBO flow working, camera outputs RGBA

+++ src/shaders/canvas_cuda.fs
@@ -1,5 +1,5 @@
 #version 330 core
-out uvec4 FragColor;
+out vec4 FragColor;
 
 //in vec3 ourColor;
 in vec2 TexCoord;
